feat: use general components in filters
- apply in tree details
- apply in hardware details

Part of #600
