--- dashboard/src/components/Cards/ConfigsList.tsx
@@ -1,59 +1,0 @@
-import { memo } from 'react';
-
-import type { IBaseCard } from '@/components/Cards/BaseCard';
-import BaseCard from '@/components/Cards/BaseCard';
-import { DumbListingContent } from '@/components/ListingContent/ListingContent';
-import ListingItem from '@/components/ListingItem/ListingItem';
-import { GroupedTestStatus } from '@/components/Status/Status';
-import type { TTreeTestsData } from '@/types/tree/TreeDetails';
-
-import FilterLink from '@/pages/TreeDetails/TreeDetailsFilterLink';
-
-interface IConfigList extends Pick<TTreeTestsData, 'configStatusCounts'> {
-  title: IBaseCard['title'];
-}
-
-const ConfigsList = ({
-  configStatusCounts,
-  title,
-}: IConfigList): JSX.Element => {
-  return (
-    <BaseCard
-      title={title}
-      content={
-        <DumbListingContent>
-          {Object.keys(configStatusCounts).map(configName => {
-            const { DONE, FAIL, ERROR, MISS, PASS, SKIP, NULL } =
-              configStatusCounts[configName];
-            return (
-              <FilterLink
-                key={configName}
-                filterSection="configs"
-                filterValue={configName}
-              >
-                <ListingItem
-                  hasBottomBorder
-                  key={configName}
-                  text={configName}
-                  leftIcon={
-                    <GroupedTestStatus
-                      done={DONE}
-                      fail={FAIL}
-                      error={ERROR}
-                      miss={MISS}
-                      pass={PASS}
-                      skip={SKIP}
-                      nullStatus={NULL}
-                    />
-                  }
-                />
-              </FilterLink>
-            );
-          })}
-        </DumbListingContent>
-      }
-    />
-  );
-};
-
-export default memo(ConfigsList);
