+++ dashboard/src/pages/TreeDetails/Tabs/Boots/BootsTab.tsx
@@ -9,24 +9,27 @@ import { useTestsTab } from '@/api/treeDetails';
 import BaseCard from '@/components/Cards/BaseCard';
 import { Skeleton } from '@/components/Skeleton';
 
-import {
-  DesktopGrid,
-  InnerMobileGrid,
-  MobileGrid,
-} from '@/pages/TreeDetails/Tabs/TabGrid';
-
 import { BootsTable } from '@/components/BootsTable/BootsTable';
-import MemoizedStatusChart from '@/components/Cards/StatusChart';
-import MemoizedConfigList from '@/components/Cards/ConfigsList';
-import MemoizedErrorsSummary from '@/components/Cards/ErrorsSummary';
 import MemoizedIssuesList from '@/components/Cards/IssuesList';
 import MemoizedHardwareTested from '@/components/Cards/HardwareTested';
 import type { TestsTableFilter } from '@/types/tree/TreeDetails';
+import {
+  DesktopGrid,
+  MobileGrid,
+  InnerMobileGrid,
+} from '@/components/Tabs/TabGrid';
+
+import MemoizedConfigList from '@/components/Tabs/Tests/ConfigsList';
+
+import MemoizedErrorsSummary from '@/components/Tabs/Tests/ErrorsSummary';
+
+import MemoizedStatusCard from '@/components/Tabs/Tests/StatusCard';
+import type { TFilter } from '@/types/general';
 
 import TreeCommitNavigationGraph from '@/pages/TreeDetails/Tabs/TreeCommitNavigationGraph';
 
 interface BootsTabProps {
-  reqFilter: Record<string, string[]>;
+  reqFilter: TFilter;
 }
 
 const BootsTab = ({ reqFilter }: BootsTabProps): JSX.Element => {
@@ -125,17 +128,19 @@ const BootsTab = ({ reqFilter }: BootsTabProps): JSX.Element => {
     <div className="flex flex-col gap-8 pt-4">
       <DesktopGrid>
         <div>
-          <MemoizedStatusChart
+          <MemoizedStatusCard
             title={<FormattedMessage id="bootsTab.bootStatus" />}
             statusCounts={data.bootStatusSummary}
           />
           <MemoizedConfigList
             title={<FormattedMessage id="bootsTab.configs" />}
             configStatusCounts={data.bootConfigs}
+            diffFilter={diffFilter}
           />
           <MemoizedErrorsSummary
             title={<FormattedMessage id="global.summary" />}
             archCompilerErrors={data.bootArchSummary}
+            diffFilter={diffFilter}
           />
           <MemoizedIssuesList
             title={<FormattedMessage id="global.issues" />}
@@ -152,7 +157,7 @@ const BootsTab = ({ reqFilter }: BootsTabProps): JSX.Element => {
         </div>
       </DesktopGrid>
       <MobileGrid>
-        <MemoizedStatusChart
+        <MemoizedStatusCard
           title={<FormattedMessage id="bootsTab.bootStatus" />}
           statusCounts={data.bootStatusSummary}
         />
@@ -162,10 +167,12 @@ const BootsTab = ({ reqFilter }: BootsTabProps): JSX.Element => {
             <MemoizedConfigList
               title={<FormattedMessage id="bootsTab.configs" />}
               configStatusCounts={data.bootConfigs}
+              diffFilter={diffFilter}
             />
             <MemoizedErrorsSummary
               title={<FormattedMessage id="global.summary" />}
               archCompilerErrors={data.bootArchSummary}
+              diffFilter={diffFilter}
             />
             <MemoizedIssuesList
               title={<FormattedMessage id="global.issues" />}
