+++ dashboard/src/pages/hardwareDetails/Tabs/Boots/BootsTab.tsx
@@ -1,41 +1,29 @@
 import { FormattedMessage } from 'react-intl';
 
-import { memo, useCallback } from 'react';
+import { useCallback } from 'react';
 
 import type { LinkProps } from '@tanstack/react-router';
 
 import { useNavigate, useSearch } from '@tanstack/react-router';
 
 import { BootsTable } from '@/components/BootsTable/BootsTable';
-import MemoizedStatusChart from '@/components/Cards/StatusChart';
-import MemoizedIssuesList from '@/components/Cards/IssuesList';
 
-import {
-  DesktopGrid,
-  InnerMobileGrid,
-  MobileGrid,
-} from '@/pages/TreeDetails/Tabs/TabGrid';
+import MemoizedIssuesList from '@/components/Cards/IssuesList';
 
 import type { THardwareDetails } from '@/types/hardware/hardwareDetails';
 
-import type {
-  TestsTableFilter,
-  TTreeTestsData,
-} from '@/types/tree/TreeDetails';
-import type { IBaseCard } from '@/components/Cards/BaseCard';
-import BaseCard from '@/components/Cards/BaseCard';
-import { DumbListingContent } from '@/components/ListingContent/ListingContent';
+import type { TestsTableFilter } from '@/types/tree/TreeDetails';
 
-import ListingItem from '@/components/ListingItem/ListingItem';
-import { GroupedTestStatus } from '@/components/Status/Status';
+import {
+  DesktopGrid,
+  MobileGrid,
+  InnerMobileGrid,
+} from '@/components/Tabs/TabGrid';
 
-import { DumbSummary } from '@/components/Summary/Summary';
+import MemoizedStatusCard from '@/components/Tabs/Tests/StatusCard';
+import MemoizedConfigList from '@/components/Tabs/Tests/ConfigsList';
+import MemoizedErrorsSummary from '@/components/Tabs/Tests/ErrorsSummary';
 
-import type { ArchCompilerStatus } from '@/types/general';
-
-import FilterLink from '@/pages/hardwareDetails/HardwareDetailsFilterLink';
-
-import { MemoizedSummaryItem } from '@/pages/hardwareDetails/Tabs/Build/BuildTab';
 import HardwareCommitNavigationGraph from '@/pages/hardwareDetails/Tabs/HardwareCommitNavigationGraph';
 
 interface TBootsTab {
@@ -44,110 +32,7 @@ interface TBootsTab {
   hardwareId: string;
 }
 
-interface IConfigList extends Pick<TTreeTestsData, 'configStatusCounts'> {
-  title: IBaseCard['title'];
-}
-
-interface IErrorsSummary {
-  archCompilerErrors: ArchCompilerStatus[];
-  title: IBaseCard['title'];
-}
-
-const ConfigsList = ({
-  configStatusCounts,
-  title,
-}: IConfigList): JSX.Element => {
-  return (
-    <BaseCard
-      title={title}
-      content={
-        <DumbListingContent>
-          {Object.keys(configStatusCounts).map(configName => {
-            const { DONE, FAIL, ERROR, MISS, PASS, SKIP, NULL } =
-              configStatusCounts[configName];
-            return (
-              <FilterLink
-                key={configName}
-                filterSection="configs"
-                filterValue={configName}
-              >
-                <ListingItem
-                  hasBottomBorder
-                  key={configName}
-                  text={configName}
-                  leftIcon={
-                    <GroupedTestStatus
-                      done={DONE}
-                      fail={FAIL}
-                      error={ERROR}
-                      miss={MISS}
-                      pass={PASS}
-                      skip={SKIP}
-                      nullStatus={NULL}
-                    />
-                  }
-                />
-              </FilterLink>
-            );
-          })}
-        </DumbListingContent>
-      }
-    />
-  );
-};
-
-//TODO: put it in other file to be reused
-export const MemoizedConfigList = memo(ConfigsList);
-
-const summaryHeaders = [
-  <FormattedMessage key="global.arch" id="global.arch" />,
-  <FormattedMessage key="global.compiler" id="global.compiler" />,
-];
-
-const ErrorsSummary = ({
-  archCompilerErrors,
-  title,
-}: IErrorsSummary): JSX.Element => {
-  return (
-    <BaseCard
-      title={title}
-      content={
-        <DumbSummary summaryHeaders={summaryHeaders}>
-          {archCompilerErrors.map(e => {
-            const statusCounts = e.status;
-            const currentCompilers = [e.compiler];
-            return (
-              <MemoizedSummaryItem
-                key={e.arch}
-                arch={{
-                  text: e.arch,
-                }}
-                leftIcon={
-                  <GroupedTestStatus
-                    forceNumber={false}
-                    done={statusCounts.DONE}
-                    fail={statusCounts.FAIL}
-                    error={statusCounts.ERROR}
-                    miss={statusCounts.MISS}
-                    pass={statusCounts.PASS}
-                    skip={statusCounts.SKIP}
-                    nullStatus={statusCounts.NULL}
-                  />
-                }
-                compilers={currentCompilers}
-              />
-            );
-          })}
-        </DumbSummary>
-      }
-    />
-  );
-};
-
-//TODO: put it in other file to be reused
-export const MemoizedErrorsSummary = memo(ErrorsSummary);
-
-const BootsTab = ({ boots, trees, hardwareId }: TBootsTab): JSX.Element => {
+const BootsTab = ({ boots, hardwareId, trees }: TBootsTab): JSX.Element => {
   const { tableFilter, diffFilter } = useSearch({
     from: '/hardware/$hardwareId',
   });
@@ -205,17 +90,21 @@ const BootsTab = ({ boots, trees, hardwareId }: TBootsTab): JSX.Element => {
     <div className="flex flex-col gap-8 pt-4">
       <DesktopGrid>
         <div>
-          <MemoizedStatusChart
+          <MemoizedStatusCard
             title={<FormattedMessage id="bootsTab.bootStatus" />}
             statusCounts={boots.statusSummary}
           />
           <MemoizedConfigList
+            disabled
             title={<FormattedMessage id="bootsTab.configs" />}
             configStatusCounts={boots.configs}
+            diffFilter={diffFilter}
           />
           <MemoizedErrorsSummary
+            disabled
             title={<FormattedMessage id="global.summary" />}
             archCompilerErrors={boots.archSummary}
+            diffFilter={diffFilter}
           />
           <MemoizedIssuesList
             title={<FormattedMessage id="global.issues" />}
@@ -226,7 +115,7 @@ const BootsTab = ({ boots, trees, hardwareId }: TBootsTab): JSX.Element => {
         <HardwareCommitNavigationGraph trees={trees} hardwareId={hardwareId} />
       </DesktopGrid>
       <MobileGrid>
-        <MemoizedStatusChart
+        <MemoizedStatusCard
           title={<FormattedMessage id="bootsTab.bootStatus" />}
           statusCounts={boots.statusSummary}
         />
@@ -236,10 +125,12 @@ const BootsTab = ({ boots, trees, hardwareId }: TBootsTab): JSX.Element => {
             <MemoizedConfigList
               title={<FormattedMessage id="bootsTab.configs" />}
               configStatusCounts={boots.configs}
+              diffFilter={diffFilter}
             />
             <MemoizedErrorsSummary
               title={<FormattedMessage id="global.summary" />}
               archCompilerErrors={boots.archSummary}
+              diffFilter={diffFilter}
             />
             <MemoizedIssuesList
               title={<FormattedMessage id="global.issues" />}
